% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\name{assign_gid}
\alias{assign_gid}
\title{clusters explanatory variables of high absolute correlation}
\usage{
assign_gid(data0, edge_cor)
}
\arguments{
\item{data0}{data.frame : explained variable (1st column), and explanatory variables (2nd-last columns) to be contracted}

\item{edge_cor}{real number : threshold for grouping; variables of correlation no less than this threshold are grouped together}
}
\value{
gid :  array of integer :  group-ids for explanatory variables in "data0"
}
\description{
This function clusters explanatory variables of high absolute correlation and assign group ids for them.
}
\examples{
data(Cars93, package = "MASS");
data=Cars93;
data=data[complete.cases(data),];
data=data[,sapply(data[1,],is.numeric)];
assign_gid(data,edge_cor=0.9);

}
\author{
Hiroshi C. Ito
}
