% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\name{check_contract_edge}
\alias{check_contract_edge}
\title{counts numbers of variables in data after contraction under different values for "edge_cor"}
\usage{
check_contract_edge(data0, start_edge_cor = 0.2, end_edge_cor = 0.7,
  ndiv = 10, edge_explain, show = TRUE)
}
\arguments{
\item{data0}{: data.frame or matrix for data.}

\item{start_edge_cor}{real number : initial value for "edge_cor".}

\item{end_edge_cor}{real number : last value for "edge_cor".}

\item{ndiv}{integer : number of examined values for "edge_cor" between "start_edge_cor" and "end_edge_cor".}

\item{edge_explain}{real number : see arguments for function "get_pcom".}

\item{show}{TRUE/FALSE :  if TRUE, result is plotted.}
}
\value{
data.frame(edge_cor,edge_explain,nvar,nvar_single)
}
\description{
This function varies "edge_cor" and counts the number of explanatry variables in "data1" that is contracted data of "data0".
}
\examples{
data(Cars93, package = "MASS");
data=Cars93;
data=data[complete.cases(data),];
data=data[,sapply(data[1,],is.numeric)];
check_contract_edge(data,start_edge_cor=0.2,end_edge_cor=0.95,ndiv=10,edge_explain=0.8,show=1)

}
\author{
Hiroshi C. Ito
}
