% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\name{check_interaction}
\alias{check_interaction}
\title{Checks interactions in the result output of "cont_selec"}
\usage{
check_interaction(res, edge_param_number = 3, edge_cor = 1,
  omit_contracted = TRUE, repeatn = 50, use_pforeach = TRUE,
  perm = TRUE, check_spa_cor = FALSE)
}
\arguments{
\item{res}{list : result output of "cont_selec".}

\item{edge_param_number}{real number: this parameter constrains the number of explanatory variables in the subset model selection and stepwise model selection, so that [sample size]/[number of free parameter] >= "edge_param_number".}

\item{edge_cor}{real number: explanatory variables in result with absolute correlations higher than "edge_cor" are printed out.}

\item{omit_contracted}{TRUE/FALSE : if TRUE, this function checks interactions among only explanatory variables that are uncontracted ones (i.e., res$group$ngid==1).}

\item{repeatn}{integer : the number of model evaluations proccessed as a single job for each CPU core (This parameter is meaningfur only when parallel proccessing is conducted by library "pforeach").}

\item{use_pforeach}{TRUE/FALSE : if TRUE, pforeach is used instead of foreach.}

\item{perm}{TRUE/FALSE : if TRUE, permutation tests for explained variables are conducted.}

\item{check_spa_cor}{TRUE/FALSE : if TRUE, the best model with residuals significantly correlated with c(pos_x, pos_y) are removed repeatedly until the correlation becomes non-significant.}

\item{check_spa_cor_nmodel}{integer : for the top [check_spa_cor_nmodel] models, spatial correlation of residuals are checked.}
}
\value{
list(target=target,result=result,data0=data0,data1=data1,group=group,
family=family,bestmodels=res_sb$bestmodels,bestmodel_stepwise=res_ss$bestmodel,
bestmodel=res_sb$bestmodel,pos_x=pos_x,pos_y=pos_y);
}
\description{
This function check interactions among statistically-strong explanatory variables in the result output of "cont_selec"
}
\details{
Details.
}
\examples{

data(Cars93, package = "MASS");
data=Cars93;
data=data[complete.cases(data),];
data=data[,sapply(data[1,],is.numeric)];
res=cont_selec(data,target="Horsepower",edge_cor=0.9,edge_explain=0.6,
               edge_param_number=3,family="gaussian",use_pforeach=FALSE);

res_int=check_interaction(res,edge_cor=0.9,use_pforeach=FALSE)
plot_each_effect(res_int)
plot_combined_effect(res_int,sign_effect=-1)
plot_combined_effect(res_int,sign_effect=1)


}
\author{
Hiroshi C. Ito
}
