% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\name{cont_selec}
\alias{cont_selec}
\title{contracts explanatory variables, conducts model selection and examins statistical strength of explanatory variables}
\usage{
cont_selec(data00, pos_x = NULL, pos_y = NULL, target = 1,
  edge_cor = 0.52, edge_explain = 0.65, edge_param_number = 3,
  repeatn = 100, family = "glmm_poisson", check_spa_cor = F,
  use_pforeach = TRUE, perm = TRUE, nrep = 10000)
}
\arguments{
\item{data00}{data.frame : an explained variable and explanatory variables. The explained variable is specified by "target".}

\item{pos_x}{array of numeric : x-coordinates of spatial position for each sampled data in "data00".}

\item{pos_y}{array of numeric : y-coordinates of spatial position for each sampled data in "data00".}

\item{target}{integer or character : spefifies explained variables by its column id (integer) or column name (character).}

\item{edge_cor}{real number : threshold for grouping; variables of correlation no less than this threshold are grouped together.}

\item{edge_explain}{real number : threshold for the variance ratio to be explained, based on which the number of principal components "np" to represent the contracted group is determined.}

\item{edge_param_number}{real number: this parameter constrains the number of explanatory variables in the subset model selection and stepwise model selection, so that [sample size]/[number of free parameter] >= "edge_param_number".}

\item{repeatn}{integer : the number of model evaluations proccessed as a single job for each CPU core (This parameter is meaningful only when "use_pforeach" is TRUE).}

\item{family}{character : error distribution. Possible choices are "glmm_poisson", "poisson","gaussian".}

\item{check_spa_cor}{TRUE/FALSE : if TRUE, the best model with residuals significantly correlated with c(pos_x, pos_y) are removed repeatedly until the correlation becomes non-significant.}

\item{use_pforeach}{TRUE/FALSE : if TRUE, pforeach is used instead of foreach in best subset selection and permutation tests (when "perm" is TRUE).}

\item{perm}{TRUE/FALSE : if TRUE, permutation tests for explained variables are conducted.}

\item{nrep}{integer : number of resampling in the permutation test.}
}
\value{
list(target=target, result=result, data0=data0, data1=data1, group=group, family=family, bestmodels=res_sb$bestmodels, bestmodel_stepwise=res_ss$bestmodel, bestmodel=res_sb$bestmodel, pos_x=pos_x, pos_y=pos_y)
}
\description{
This function contracts explanatory variables and conduct best subset model selection, permutation test and stepwise model selection.
}
\details{
Spatial correlation of residuals of each fitted model can also be examimed.
}
\examples{

data(Cars93, package = "MASS");
data=Cars93;
data=data[complete.cases(data),];
data=data[,sapply(data[1,],is.numeric)];

res=cont_selec(data,target="Horsepower",edge_cor=0.9,edge_explain=0.6,
               edge_param_number=3,family="gaussian",use_pforeach=FALSE);

plot_each_effect(res)
plot_combined_effect(res,sign_effect=-1)
plot_combined_effect(res,sign_effect=1)

print(res);
summary(res);
}
\author{
Hiroshi C. Ito
}
