% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\docType{package}
\name{contselec-package}
\alias{contselec-package}
\alias{_PACKAGE}
\alias{contselec}
\title{Model selection procesure by contracting explanatry variables of high absolute correlation, with examining statistical strength of each explanatry variable.}
\description{
Model selection procesure by contracting explanatry variables of high absolute correlation, with examining statistical strength of each explanatry variable. Main function is "cont_selec", which consists of "contract" (grouping explanatory variables of high absolute correlation, and contracting them by using theprincipal components of the groups) and "bestsub_stepwise" (combination of best subset selection, permutation test and stepwise selection)
}
\author{
\strong{Maintainer}: Hiroshi C. Ito \email{hiroshibeetle@gmail.com}

}
\keyword{internal}
