% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\name{get_pcom}
\alias{get_pcom}
\title{gets PCA scores from data subset}
\usage{
get_pcom(data0s, edge_explain, np = -1, scale = FALSE)
}
\arguments{
\item{data0s}{: data.frame or matrix}

\item{edge_explain}{real number : "np" is chosen so that [explained variance]/[total variance] > "edge_explain", when "np" is not given}

\item{np}{integer : number of principal components extracted}

\item{scale}{TRUE/FALSE : if TRUE, variables are scaled in advance so that mean=0 and sd=1.}
}
\value{
list(np, data0s_pcom), where "np" (integer) is number of principal components, and "data0s_pcom" (matrix) contains PCA scores for the corresponding principal components.
}
\description{
This function returns PCA scores for the 1th,...,"np"th principal components of "data0s" so that [explained variance]/[total variance] is larger than "edge_explain"
}
\examples{
data(Cars93, package = "MASS");
data=Cars93
data=data[complete.cases(data),];
data=data[,sapply(data[1,],is.numeric)];
gid=assign_gid(data,edge_cor=0.8);
print(gid);
pcom=get_pcom(data[,gid==2],edge_explain=0.85,scale=TRUE);
}
\author{
Hiroshi C. Ito
}
