% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\docType{data}
\name{param0}
\alias{param0}
\title{Default parameter set for "plot_each_effect" and "plot_combined_effect".}
\format{An object of class \code{list} of length 23.}
\usage{
param0
}
\description{
This list is a default parameter set for "plot_each_effect" and "plot_combined_effect" to plot result output of "cont_selec", "selec_bestsub_stepwise" and "check_interaction".
}
\author{
Hiroshi C. Ito
}
\keyword{datasets}
