% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\name{plot_combined_effect}
\alias{plot_combined_effect}
\title{plots combined effect of statistically-strong explanatory variables}
\usage{
plot_combined_effect(res, param = param0, outfile = "testout",
  outpng = F, outeps = F, outpng_show = T, hide_title = F,
  sign_effect = 1, do_order = TRUE)
}
\arguments{
\item{res}{list : result output of "cont_selec" or "selec_bestsub_stepwise"}

\item{param}{list : graphics parameter. Default values are given by "param0"}

\item{outfile}{character : filename of image output.}

\item{outpng}{TRUE/FALSE : if TRUE, plottings are saved in png format.}

\item{outeps}{TRUE/FALSE : if TRUE, plottings are saved in eps format.}

\item{outpng_show}{TRUE/FALSE : if TRUE, saved png files are shown.}

\item{hide_title}{TRUE/FALSE : if TRUE, titles of plottings are hidden.}

\item{sign_effect}{1/-1 : if 1, only explanatory variables with positive regression coeffecients are shown, and vice versa.}

\item{do_order}{TRUE/FALSE : if TRUE, ponds are ordered along horisontal axis in terms of maximum combined impacts.}
}
\description{
This function plots the expected combined effect of statistically-strong explanatory variables detected by "cont_selec" or "selec_bestsub_stepwise". See "cont_selec" in this manual or help(cont_selec).
}
\author{
Hiroshi C. Ito
}
