% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\name{plot_each_effect}
\alias{plot_each_effect}
\title{plots effects of statistically-strong explanatory variables separately}
\usage{
plot_each_effect(res, param = param0, outfile = "testout",
  outpng = F, outeps = F, outpng_show = T, hide_title = F,
  sign_flip = FALSE, pch = 24)
}
\arguments{
\item{res}{list : result output of "cont_selec"  or "selec_bestsub_stepwise"}

\item{param}{list : graphics parameter. Default values are given by "param0"}

\item{outfile}{character : filename of image output.}

\item{outpng}{TRUE/FALSE : if TRUE, plottings are saved in png format.}

\item{outeps}{TRUE/FALSE : if TRUE, plottings are saved in eps format.}

\item{outpng_show}{TRUE/FALSE : if TRUE, saved png files are shown.}

\item{hide_title}{TRUE/FALSE : if TRUE, titles of plottings are not shown.}

\item{sign_flip}{TRUE/FALSE : if TRUE, signs of mean and max impacts of explanatory variables are flipped.}

\item{pch}{integer or "text" : pch for schatter plot. If "text", then sample ids are plotted.}
}
\description{
This function plots separately the expected effect of each statistically-strong explanatory variable detected by "cont_selec" or "selec_bestsub_stepwise". See "cont_selec" in this manual or help(cont_selec).
}
\author{
Hiroshi C. Ito
}
