% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\name{pngout}
\alias{pngout}
\title{saves plotting in png format}
\usage{
pngout(dev_id = as.numeric(dev.list()[length(dev.list())]),
  plotfile = "testout", density = 150, geometry = 100 * 100/density,
  outeps = FALSE, prefix = "./", show = TRUE,
  font = param0$font_family, outpng = TRUE)
}
\arguments{
\item{dev_id}{integer : device id (default value is current device id)}

\item{plotfile}{character : filename}

\item{density}{integer : "density" option for "convert"}

\item{geometry}{integer : "geometry" option for "convert"}

\item{outeps}{TRUE/FALSE : if TRUE plotting is saved in eps format}

\item{prefix}{character : location of saved file}

\item{show}{TRUE/FALSE : if TRUE saved png file is shown by "display" command.}

\item{font}{character : font family}

\item{outpng}{TRUE/FALSE : if TRUE plotting is saved in eps format}
}
\description{
This function saves plotting in png and/or eps format by using "dev.copy2eps" and command "convert" in ImageMagick.
}
\author{
Hiroshi C. Ito
}
