% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\name{put_sigmark}
\alias{put_sigmark}
\title{puts sigmarks}
\usage{
put_sigmark(result)
}
\arguments{
\item{result}{: output of "selec_bestsub_stepwise"}
}
\description{
This function puts sigmarks "*","**" and "***" for p-values <0.05, <0.01 and <0.001 to the result output of "selec_stepwise".
}
\author{
Hiroshi C. Ito
}
