% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\name{selec_bestsub}
\alias{selec_bestsub}
\title{best-subset model selection}
\usage{
selec_bestsub(data0, data1, group, pos_x = NULL, pos_y = NULL,
  edge_param_number = 3, repeatn = 100, family,
  check_spa_cor = FALSE, target = "", use_pforeach = TRUE,
  perm = TRUE, daic = 2, check_spa_cor_nmodel = 50, nrep = 10000)
}
\arguments{
\item{data0}{data.frame : uncontracted data}

\item{data1}{data.frame : constracted data}

\item{group}{(list(gid,gid_data1,ngid,vname1_orig,vname0_orig)) : information about contraction of "data0" into "data1".}

\item{pos_x}{array of numeric : x-coordinates of spatial position for each sampled data in "data0".}

\item{pos_y}{array of numeric : y-coordinates of spatial position for each sampled data in "data0".}

\item{edge_param_number}{real number: this parameter constrains the number of explanatory variables in the subset model selection and stepwise model selection, so that [sample size]/[number of free parameter] >= "edge_param_number".}

\item{repeatn}{integer : the number of model evaluations proccessed as a single job for each CPU core (This parameter is meaningfur only when parallel proccessing is conducted by library "pforeach").}

\item{family}{character : error distribution. Possible choices are "glmm_poisson", "poisson","gaussian".}

\item{check_spa_cor}{TRUE/FALSE : if TRUE, the best model with residuals significantly correlated with c(pos_x, pos_y) are removed repeatedly until the correlation becomes non-significant.}

\item{target}{character : arbitrary name for explained variable "y".}

\item{use_pforeach}{TRUE/FALSE : if TRUE, pforeach is used instead of foreach.}

\item{perm}{TRUE/FALSE : if TRUE, permutation tests for explained variables are conducted.}

\item{daic}{numeric : models with Delta-AIC no more than "daic" are examined. Default value is 2.0.}

\item{nrep}{integer : number of resampling in the permutation test.}
}
\value{
list(result=result,bestmodels=aa,bestmodel=bestmodel,
n_model=n_model,n_model_total=n_model_total)
}
\description{
This function conducts best-subset model selection
}
\details{
Details.
}
\examples{
data(Cars93, package = "MASS");
data=Cars93;
data=data[complete.cases(data),];
data=data[,sapply(data[1,],is.numeric)];
con=contract(data,edge_cor=0.9,edge_explain=0.6,target="Horsepower");
data1=con$data1;
for(i in 2:ncol(data1))data1[,i]=(data1[,i]-min(data1[,i]))/(max(data1[,i])-min(data1[,i]));

res=selec_bestsub(con$data0,data1,con$group,edge_param_number=3,repeatn=50,family="gaussian",target=con$target, use_pforeach=FALSE,perm=TRUE,daic=2.0);

print(res$result);



}
\author{
Hiroshi C. Ito
}
