% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contselec-package8.R
\name{summary.selec_bestsub_stepwise}
\alias{summary.selec_bestsub_stepwise}
\title{shows summary of "cont_selec" or "selec_bestsub_stepwise" result.}
\usage{
\method{summary}{selec_bestsub_stepwise}(res)
}
\arguments{
\item{res}{: output of "selec_bestsub_stepwise"}
}
\description{
This function shows summary of the result of "cont_selec" or "selec_bestsub_stepwise". See "cont_selec" for an example.
}
\author{
Hiroshi C. Ito
}
